Removed HexCoords.Elevation
Including elevation as part of the HexCoords class was a mistake since HexCoords are used for the key for our hastable of tiles, but since you can never have two tiles occupying the same space with differing elevation, that was being ignored in certain situations but not others. Basically, I think it makes more sense to have the elevation stored only by the value in the hashtable (the HexTile object) rather than the key (the HexCoords).