Fixed bug where FindChunkForCoordinates would sometimes get the wrong chunk
FindChunkForCoordinates previous returned the first chunk that concevably could contain a tile at the specified coordinates, even if none existed. This is an issue because chunks with different materials frequently have overlapping bounds, so the method wouldn't always return the chunk that actually contained the tile at this position. Now, it will return the chunk containing the tile at the specified position or null (even if there is a chunk with bounds that cover the specified coordinates but there is no tile there).